device/requirements.txt
boto3
joblib
scikit-learn
awsiotsdk
pandas
